case sensitive test

+++ src/scorer.rs
@@ -70,13 +70,13 @@ pub fn jaro_similarity(
     let m = word2.len();
     let word1_chars = char_vec(word1, case_sensitive, remove_whitespace);
     let word2_chars = char_vec(word2, case_sensitive, remove_whitespace);
-    let max_dist = (usize::max(m, n) / 2) - 1;
+    let max_dist = (i32::max(m as i32, n as i32) / 2) - 1;
     let mut matches = 0;
     let mut hash_word1 = vec![0; n];
     let mut hash_word2 = vec![0; m];
     for i in 0..n {
-        let mut j = i32::max(i as i32 - max_dist as i32, 0);
-        while j < usize::min(i + max_dist + 1, m) as i32 {
+        let mut j = i32::max(i as i32 - max_dist, 0);
+        while j < usize::min(i + max_dist as usize + 1, m) as i32 {
             if word1_chars[i] == word2_chars[j as usize] && hash_word2[j as usize] == 0 {
                 hash_word1[i] = 1;
                 hash_word2[j as usize] = 1;
